requirements.txt
faiss_cpu==1.11.0.post1
fastapi==0.116.1
numpy==2.3.2
openai==1.97.1
pandas==2.3.1
pydantic==2.11.7
pyspark==3.5.5
python-dotenv==1.1.1
Requests==2.32.4
scikit_learn==1.7.1
sentence_transformers==5.0.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
streamlit==1.47.1
plotly==6.2.0
matplotlib==3.10.3